create: add v6_connect_timeout and resolve_retries
DNS resolution may fail, by default, retry resolving 3 times (when there's no
resolution within resolve_timeout).

The connect_timeout is used for all connection attempts to the last destination.
If there is a connection attempt to an IPv6 address (and others are remaining),
use v6_connect_timeout (default 200ms) instead. This is the same value as curl.
In some network environments, the previously used connect_timeout of one second
was not sufficient to establish a complete connection.